Revert "feat: Threads"
