+++ src/gateway/handlers/mod.ts
@@ -68,16 +68,6 @@ import { applicationCommandCreate } from './applicationCommandCreate.ts'
 import { applicationCommandDelete } from './applicationCommandDelete.ts'
 import { applicationCommandUpdate } from './applicationCommandUpdate.ts'
 import type { SlashCommand } from '../../interactions/slashCommand.ts'
-import type {
-  ThreadChannel,
-  ThreadMember
-} from '../../structures/threadChannel.ts'
-import { threadCreate } from './threadCreate.ts'
-import { threadDelete } from './threadDelete.ts'
-import { threadUpdate } from './threadUpdate.ts'
-import { threadMembersUpdate } from './threadMembersUpdate.ts'
-import { threadMemberUpdate } from './threadMemberUpdate.ts'
-import { threadListSync } from './threadListSync.ts'
 
 export const gatewayHandlers: {
   [eventCode in GatewayEvents]: GatewayEventHandler | undefined
@@ -122,13 +112,7 @@ export const gatewayHandlers: {
   VOICE_STATE_UPDATE: voiceStateUpdate,
   VOICE_SERVER_UPDATE: voiceServerUpdate,
   WEBHOOKS_UPDATE: webhooksUpdate,
-  INTERACTION_CREATE: interactionCreate,
-  THREAD_CREATE: threadCreate,
-  THREAD_DELETE: threadDelete,
-  THREAD_UPDATE: threadUpdate,
-  THREAD_LIST_SYNC: threadListSync,
-  THREAD_MEMBERS_UPDATE: threadMembersUpdate,
-  THREAD_MEMBER_UPDATE: threadMemberUpdate
+  INTERACTION_CREATE: interactionCreate
 }
 
 export interface VoiceServerUpdateData {
@@ -445,25 +429,4 @@ export type ClientEvents = {
   commandUsed: [ctx: CommandContext]
   commandError: [ctx: CommandContext, err: Error]
   gatewayError: [err: ErrorEvent, shards: [number, number]]
-
-  threadCreate: [thread: ThreadChannel]
-  threadDelete: [thread: ThreadChannel]
-  threadDeleteUncached: [thread: string]
-  threadUpdate: [old: ThreadChannel, new: ThreadChannel]
-  threadUpdateUncached: [thread: ThreadChannel]
-  threadListSync: [
-    guild: Guild,
-    threads: Collection<string, ThreadChannel>,
-    members: Collection<string, ThreadMember>,
-    channels: Collection<string, GuildTextBasedChannel>
-  ]
-  threadMemberUpdate: [me: ThreadMember]
-  threadMembersUpdate: [
-    guild: Guild,
-    added: Collection<string, ThreadMember>,
-    removed: Collection<string, ThreadMember>,
-    memberCount: number
-  ]
-  threadMemberAdd: [member: ThreadMember, guild: Guild]
-  threadMemberRemove: [member: ThreadMember, guild: Guild]
 }
