+++ src/managers/baseChild.ts
@@ -13,10 +13,6 @@ export class BaseChildManager<T, T2> extends Base {
     this.parent = parent
   }
 
-  get cacheName(): string {
-    return this.parent.cacheName
-  }
-
   async get(key: string): Promise<T2 | undefined> {
     return this.parent.get(key)
   }
@@ -25,7 +21,7 @@ export class BaseChildManager<T, T2> extends Base {
     return this.parent.set(key, value)
   }
 
-  async delete(_: string): Promise<any> {
+  async delete(key: string): Promise<any> {
     return false
   }
 
@@ -73,6 +69,6 @@ export class BaseChildManager<T, T2> extends Base {
   }
 
   [Deno.customInspect](): string {
-    return `ChildManager(${this.cacheName})`
+    return `ChildManager(${this.parent.cacheName})`
   }
 }
