--- src/structures/resolvable.ts
@@ -1,6 +1,0 @@
-export interface IResolvable<T> {
-  id: string
-  get: () => Promise<T | undefined>
-  fetch: () => Promise<T>
-  resolve: () => Promise<T | undefined>
-}
