+++ src/types/constants.ts
@@ -2,6 +2,6 @@ export const Constants = {
   DISCORD_API_URL: 'https://discord.com/api',
   DISCORD_GATEWAY_URL: 'wss://gateway.discord.gg',
   DISCORD_CDN_URL: 'https://cdn.discordapp.com',
-  DISCORD_API_VERSION: 9,
+  DISCORD_API_VERSION: 8,
   DISCORD_VOICE_VERSION: 4
 }
