+++ src/types/guild.ts
@@ -16,8 +16,7 @@ import type {
   GuildNewsChannelPayload,
   GuildTextBasedChannelPayload,
   GuildTextChannelPayload,
-  GuildVoiceChannelPayload,
-  ThreadChannelPayload
+  GuildVoiceChannelPayload
 } from './channel.ts'
 import type { EmojiPayload } from './emoji.ts'
 import type { PresenceUpdatePayload } from './gateway.ts'
@@ -75,7 +74,6 @@ export interface GuildPayload {
   approximate_number_count?: number
   approximate_presence_count?: number
   nsfw: boolean
-  threads?: ThreadChannelPayload[]
 }
 
 export interface MemberPayload {
